Move FossaV1 and ProjectInference out of the Analyze hierarchy
