+++ spectrometer.cabal
@@ -90,13 +90,13 @@ library
   -- cabal-fmt: expand src
   exposed-modules:
     App.Fossa.Analyze
-    App.Fossa.Analyze.FossaV1
     App.Fossa.Analyze.Graph
     App.Fossa.Analyze.GraphBuilder
     App.Fossa.Analyze.GraphMangler
     App.Fossa.Analyze.Project
-    App.Fossa.Analyze.ProjectInference
+    App.Fossa.FossaAPIV1
     App.Fossa.Main
+    App.Fossa.ProjectInference
     App.Fossa.Test
     App.Pathfinder.Main
     App.Pathfinder.Scan
